[38][Refactor] Change name of Folder weatherIcon => weather-icon
