+++ src/ui/widget/weather-icon/weather-icon.tsx
