+++ src/ui/widget/widget.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { IWeather } from '../../common/types/IWeather';
 import Temperature from './temperature/temperature';
 import City from './city/city';
-import WeatherIcon from './weatherIcon/weather-icon';
+import WeatherIcon from './weather-icon/weather-icon';
 import { styled, Typography } from '@mui/material';
 import { temperatureFontSize } from '../../common/styles/consts';
 import DateContainer from './date-container/date-container';
